moved some files, fixed placement of Sancturary chest
